Refactor link box, add new edit modal
